feat(engine): a channel can hear another channel's oscillator
Steps 03 and 04 of `docs/plans/typed-audio-edges/`, and the acceptance
measurements from 05 that do not need a face. The compiler landed in the
last two commits and handed out tap indices; nothing allocated one, no
producer wrote anywhere, and the channel loop still walked `0..active`.

**The taps.** A producer is handed a port group for the duration of one
process call -- `AudioTaps`, indexed by the device's own tap number, which
is its outlet's position in the declared audio run. Nothing is retained
across calls, which is `AUDIO_ARCHITECTURE.md`'s rule for auxiliary
buffers. ML-P8 fills seven and DS-01 four, at the points their descriptors
declare, summed across active voices and panned with the voice.

A subscriber is one more reason to *run* a source, folded into the same
decision the patch already makes: `Osc 3` turned down to silence is
exactly the outlet somebody else wants, and without this the acceptance
case would have published the zeros of an oscillator that was skipped.

**The order.** The channel loop walks `graph.order()`. A muted producer
somebody reads renders its generator and stops there -- mute is an
output-stage decision about what reaches the bus, and a pre-level tap is
the signal a muted source still has -- while a muted producer nobody reads
still skips. The modulator tick pass stays in index order and stays
separate, so no subscription can move another channel's LFO phase.

**Aux In.** `DeviceKind::AuxIn`, three descriptor-addressed parameters,
serialized as `aux_in`. It publishes its own output so Aux Ins can chain,
which is also what makes a ring constructible and the cycle refusal
reachable from the program rather than only from a unit test. A
subscription is a channel-scoped address and moves through the same
rescope pass as a route's destination; a deleted producer sends it to
`DEPARTED_SOURCE` rather than to whichever channel inherited the index.

`ParamCurve::Stepped` widened from `u8` to `u16`: a selector that names a
channel needs 257 positions, and the byte was the default integer rather
than a designed ceiling.

Ten engine tests, and both halves of the headline one: an Aux In hears a
`Osc 3` that ML-P8's own output does not carry, measured spectrally in
both directions. Same at 128 and 512 frames, same offline as live, a ring
refused and given back when broken, a muted producer publishing, and a
project with no edges costing 4 KiB across sixteen channels and no
buffers at all.

The face is the next commit; nothing above is reachable from the UI yet.

+++ crates/mooloop-core/src/effect.rs
@@ -155,7 +155,14 @@ pub enum ParamCurve {
     /// Requires `min > 0`.
     Exponential,
     /// `n` discrete positions mapped across the range: mode selectors.
-    Stepped(u8),
+    ///
+    /// Wider than a `u8` because a selector may name a channel, and a
+    /// project may hold [`crate::MAX_CHANNELS`] of them plus the "none"
+    /// position -- 257 positions, which is two more than a byte of *count*
+    /// can express. The width was never a design decision; it was the
+    /// default integer, and the first selector that needed more than 255
+    /// positions is the one that says so.
+    Stepped(u16),
 }
 
 /// One parameter's identity, range, and mapping. The single source of truth:
